Modal added, toggle context finished, UI cleanup
Completed context allowing a user to toggle between a vertical and horizontal view; Functionality not yet implemented.

Added a modal component. Currently, it is used for adding a new ticket to an open project

Added AddTicketForm, which will handle the form and options for creating  new ticket

UI fixes so the app behaved nicer with the other components on the screen