App now pulls allTickets and list of usernames
We pull allTickets in order to know if the ticket being created is even valid. It is also used when creating a ticket to get the list of shortnames.

We pull a list of all usernames so that we can assign tickets to users that exist in our database

Fixed a bug with the InputField where the first word was always capitalized