Update to moving tickets around
Updated the logic; The actual functionality that reorders the indexes is located in Util.tsx. The old logic will be left as comments while testing, but the new code is much more streamlined.

Server is also updated so that the order is updated correctly, and the ticket status is also adjusted