Massive overhaul to ticket structure
Ran into an issue where when moving a ticket around, the app would get confused due to the component providing an index based off how many tickets where in a column, rather than its overall place in the app.

This lead to the structure being broken down into more detail. Now, the 'tickets' portion of a Project is broken down with the statuses available. This makes it easier for us to use our drag and drop, while also providing bigger clarity.

Doing this however required restructuring of our Schema and Models, as well as the certain sections of that app that dealt with the old ticket object.